+++ src/app/CMakeLists.txt
@@ -1,6 +1,8 @@
 add_executable(cmdl-args)
 
 target_sources(cmdl-args PRIVATE    app.cpp 
+                                    process.cpp
+                                    process.h
                                     parser.cpp 
                                     parser.h
                                     generator.cpp
@@ -8,6 +10,12 @@ target_sources(cmdl-args PRIVATE    app.cpp
                                     tests.h)
 
 target_include_directories(cmdl-args PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+
+set_target_properties(cmdl-args PROPERTIES VERSION 0.1.0)
+
+
+get_target_property(VERSION cmdl-args VERSION)
+target_compile_definitions(cmdl-args PRIVATE TOOL_VERSION="${VERSION}")
 
 # basic tests
